backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
redis==5.2.1
scikit-learn>=1.4.0
xgboost>=2.0.0
nltk>=3.8.0
pandas>=2.0.0
reportlab>=4.0.0
joblib>=1.3.0
requests>=2.31.0
